created duplicate computer.md under data-science
Advanced knowledge on GPU palatalization is often necessary for a Data Scientist to train his models faster on a GPU(s). That's why I created a duplicate of the original computer.md file under data-science directory as well. Although I'll be reviewing it further and look for a better alternative to it at a later point of time.